quick updates

+++ Events/Message/botMentions.js
@@ -32,25 +32,24 @@ module.exports = {
 					.setURL('https://github.com/SkullGaming31/skulledbotDiscord/issues')
 					.setLabel('Report Bugs/issues')
 			);
-
 			return message.reply({ embeds: [embed], components: [row] }).then(msg =>
 				setTimeout(() => {
 					msg.delete().catch((err) => {
 						if (err.code !== 10008) return console.error(err);
 					});
 				}, ms('1m')));
-		} else {
+
+		} else if (process.env.NODE_ENV === 'production') {
 			const row = new ActionRowBuilder().addComponents(// SkulledbotV2
 				new ButtonBuilder()
 					.setStyle(ButtonStyle.Link)
-					.setURL('https://discord.com/api/oauth2/authorize?client_id=910935752219701258&permissions=1426936031350&redirect_uri=https%3A%2F%2Fskulledbotv2.up.railway.app%2Fapi%2Fauth%2Fdiscord%2Fredirect&response_type=code&scope=identify%20connections%20guilds%20bot%20applications.commands')
+					.setURL('https://discord.com/api/oauth2/authorize?client_id=910935752219701258&permissions=1426966441078&redirect_uri=https%3A%2F%2Fskulledbotv2.up.railway.app%2Fapi%2Fauth%2Fdiscord%2Fredirect&response_type=code&scope=identify%20connections%20guilds%20bot%20applications.commands')
 					.setLabel('Invite Me'),
 				new ButtonBuilder()
 					.setStyle(ButtonStyle.Link)
 					.setURL('https://github.com/SkullGaming31/skulledbotDiscord/issues')
 					.setLabel('Report Bugs/issues')
 			);
-
 			return message.reply({ embeds: [embed], components: [row] }).then(msg =>
 				setTimeout(() => {
 					msg.delete().catch((err) => {
